Separates Data Class , change name the YelpSearResult instead Yelp,Business instead Businesses
