+++ app/src/main/java/com/example/yelpapp/model/Category.kt
@@ -1,0 +1,7 @@
+package com.example.yelpapp.model
+
+
+data class Category(
+    val alias: String,
+    val title: String
+)
