+++ app/src/main/java/com/example/yelpapp/model/Center.kt
@@ -1,0 +1,6 @@
+package com.example.yelpapp.model
+
+data class Center(
+    val latitude: Double,
+    val longitude: Double
+)
