+++ app/src/main/java/com/example/yelpapp/model/Coordinates.kt
@@ -1,0 +1,6 @@
+package com.example.yelpapp.model
+
+data class Coordinates(
+    val latitude: Double,
+    val longitude: Double
+)
