+++ app/src/main/java/com/example/yelpapp/model/Region.kt
@@ -1,0 +1,5 @@
+package com.example.yelpapp.model
+
+data class Region(
+    val center: Center
+)
