+++ app/src/main/java/com/example/yelpapp/model/YelpDbService.kt
@@ -3,8 +3,8 @@ package com.example.yelpapp.model
 
 import retrofit2.Call
 import retrofit2.http.GET
-import retrofit2.http.Header
 import retrofit2.http.Query
+import java.util.concurrent.Flow
 
 interface YelpDbService {
 
@@ -12,11 +12,10 @@ interface YelpDbService {
 
         @GET("/v3/businesses/search")
         fun getYelpSearch(
-            @Header("Authorization") api: String,
             @Query("term") term: String,
             @Query("latitude") latitude: Double,
             @Query("longitude") longitude: Double
-        ): Call<Yelp>
+        ): YelpSearchResult
     }
