+++ app/src/main/java/com/example/yelpapp/model/YelpSearchResult.kt
@@ -1,0 +1,7 @@
+package com.example.yelpapp.model
+
+data class YelpSearchResult(
+    val businesses: List<Business>,
+    val region: Region,
+    val total: Int
+)
